724. Change XrmQuark's from int's to long's in support of 64-bit       architectures.  Affects lib{X11,Xaw,Xmu,Xt,dpstk}, XPrint extension,       xf86cfg, appres, editres, xedit and xman.  Recompilation of 64-bit X       applications is recommended.  Recompiling is required for 64-bit       big-endian platforms, but 64-bit little-endian platforms can probably       survive without it (Marc La France).

+++ programs/Xserver/Xprint/Xresource.h
@@ -1,4 +1,4 @@
-/* $XFree86: $ */
+/* $XFree86: xc/programs/Xserver/Xprint/Xresource.h,v 1.3 1999/12/16 02:26:25 robin Exp $ */
 
 /* $XConsortium: Xresource.h /main/1 1996/09/28 16:59:12 rws $ */
 /*
@@ -90,7 +90,7 @@ extern char *Xpermalloc(
  *
  ****************************************************************/
 
-typedef int     XrmQuark, *XrmQuarkList;
+typedef long    XrmQuark, *XrmQuarkList;
 #define NULLQUARK ((XrmQuark) 0)
 
 typedef char *XrmString;
